requirements.txt
dj-rest-auth
Django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
grpcio
httplib2
pyasn1-modules
PyJWT
pyparsing
python-dateutil
python-decouple
tz
requests
sqlparse
uritemplate
urllib3
django-cors-headers
django-email-verification
oauthlib
Pillow
pandas
dj_database_url
django_heroku
gunicorn
psycopg2
whitenoise
